Use Jolt Physics by default in newly created projects
This also adds infrastructure to make specific settings the default, but only
in projects created after this point (so that existing projects
are not affected). This can be used for progressive upgrades in the future.

This applies to both the project creation dialog and creating projects
through the command line using `touch project.godot`.