Add Android instrumented tests to the `app` module
This builds on the work from @dsnopek in https://github.com/dsnopek/javaclasswrapper-test, by importing the set of JavaClassWrapper tests from that repo within the Godot core repo in order to bootstrap and standardize how we write and run Android instrumented tests.
The approach used here should serve as a base to build upon to expand the set of instrumented tests used to validate the project's stability.

Co-authored-by: David Snopek <dsnopek@gmail.com>